+++ src/main/java/com/kakaotech/team14backend/post/application/FindPopularPosts.java
@@ -1,19 +1,20 @@
-package com.kakaotech.team14backend.inner.post.usecase;
+package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.common.RedisKey;
-import com.kakaotech.team14backend.inner.post.model.LevelIndexes;
-import com.kakaotech.team14backend.post.domain.Post;
-import com.kakaotech.team14backend.inner.post.model.PostRandomFetcher;
-import com.kakaotech.team14backend.inner.post.model.RandomIndexes;
-import com.kakaotech.team14backend.inner.post.repository.PostRepository;
 import com.kakaotech.team14backend.outer.post.dto.GetIncompletePopularPostDTO;
 import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListRequestDTO;
 import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListResponseDTO;
 import com.kakaotech.team14backend.outer.post.mapper.PostMapper;
+import com.kakaotech.team14backend.post.domain.LevelIndexes;
+import com.kakaotech.team14backend.post.domain.Post;
+import com.kakaotech.team14backend.post.domain.PostRandomFetcher;
+import com.kakaotech.team14backend.post.domain.RandomIndexes;
+import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
