+++ src/main/java/com/kakaotech/team14backend/post/application/SavePopularPosts.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.post.usecase;
+package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.common.RedisKey;
 import com.kakaotech.team14backend.outer.post.dto.GetIncompletePopularPostDTO;
