+++ src/main/java/com/kakaotech/team14backend/post/application/UpdatePostPopularity.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.post.usecase;
+package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
